# static/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation of infoblog API.",
    "title": "infoblog.",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
    "/auth/checkcode": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Проверка смс кода.",
        "operationId": "checkCodeRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CheckCodeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/checkCodeResponse"
          }
        }
      }
    },
...
    "/auth/email/registration": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Проверка смс кода.",
        "operationId": "EmailActivationRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/EmailActivationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/emailRegistrationResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CheckCodeRequest": {
      "type": "object",
      "properties": {
        "Code": {
          "type": "integer",
          "format": "int64"
        },
        "Phone": {
          "type": "string"
        }
      },
      "x-go-package": "gitlab.com/InfoBlogFriends/server/request"
    },
...
    "EmailActivationRequest": {
      "type": "object",
      "properties": {
        "Email": {
          "type": "string"
        },
        "Password": {
          "type": "string"
        }
      },
      "x-go-package": "gitlab.com/InfoBlogFriends/server/request"
    },
...
    "Response": {
      "type": "object",
      "properties": {
        "Data": {
          "type": "object"
        },
        "Msg": {
          "type": "string"
        },
        "Success": {
          "type": "boolean"
        }
      },
      "x-go-package": "gitlab.com/InfoBlogFriends/server/request"
    },
...
  },
  "responses": {
    "checkCodeResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Response"
      }
    },
    "emailRegistrationResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Response"
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}